// src/shared/dataprovider/api/model/clientDTO.ts
/**
 * DynastyGateway API
 * API Connections to DPay
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProviderDTO } from './providerDTO';
import { StatusEnum } from './statusEnum';

export interface ClientDTO { 
    Id?: number;
    Name?: string;
    TransactionCount?: number;
    RemittanceWalletBalance?: number;
    TotalWalletWithdrawn?: number;
    TotalCashIn?: number;
    TotalCashOut?: number;
    Balance?: number;
    TopUpBalance?: number;
    Status?: StatusEnum;
    CreatedDate?: Date;
    LastUpdatedDate?: Date;
    ProviderId?: number;
    CallbackUrl?: string;
    RedirectUrl?: string;
    AccessKey?: string;
    SecretKey?: string;
    Provider?: ProviderDTO;
}